modules/location-ingester/requirements.txt
kafka-python==2.0.2
grpcio-tools==1.51.3
 